backend/govimithurubackend/claimManagement/serializers.py
from rest_framework import serializers

from .models import Claim, Evidence

class ClaimSerializer(serializers.ModelSerializer):
    class Meta:
        model = Claim
        fields = '__all__'

class EvidenceSerializer(serializers.ModelSerializer):
    class Meta:
        model = Evidence
        fields = '__all__'