backend/govimithurubackend/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
schedule==1.1.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
whitenoise==5.3.0
Pillow==8.4.0
